Web login

+++ databus/web/app.py
@@ -5,9 +5,10 @@ Theme help: https://bootswatch.com/darkly/
 from datetime import datetime
 import io
 import uuid
-from flask import Flask, redirect, render_template, request, send_file, url_for
+from flask import Flask, redirect, render_template, request, send_file, session, url_for
 from waitress import serve
 import databus
+from databus.client.client import Client, Credential
 from databus.dispatcher.abstract_dispatcher import AbstractDispatcher
 from databus.passenger.attachment import Attachment, AttachmentFormat
 from databus.pqueue.queue_status import QueueStatus
@@ -15,12 +16,13 @@ from databus.report.client_log import ClientLogReader
 from databus.report.client_queue import ClientPassengerQueueReader
 from databus.report.puller_peek import PullerPeek
 
-
 ##############################
 # Main stuff
 ##############################
 
 _APP = Flask(__name__)
+_APP.secret_key = "databus"
+_APP.config["CACHE_TYPE"] = "null"
 _DISPATCHER: AbstractDispatcher
 
 
@@ -35,9 +37,12 @@ def run_web_server(dispatcher: AbstractDispatcher):
 # Home page
 ##############################
 
-@_APP.route("/")
+@_APP.route("/home")
 def _home():
-    global _DISPATCHER # pylint: disable=W0603
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+
     return render_template("home.html")
 
 ##############################
@@ -47,13 +52,27 @@ def _home():
 @_APP.route("/log_list")
 def _log_list():
     global _DISPATCHER # pylint: disable=W0603
-    entries = ClientLogReader(_DISPATCHER).get_client_log_list()
+
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+    if authenticated_client_id == Client.ROOT:
+        authenticated_client_id = None
+
+    entries = ClientLogReader(_DISPATCHER).get_client_log_list(p_client_id=authenticated_client_id)
     return render_template("log_list.html", entries=entries)
 
 
 @_APP.route("/log_display")
 def _log_display():
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+
     client_id = request.args.get("client", 0, type=str)
+    if authenticated_client_id not in (Client.ROOT, client_id):
+        return redirect(url_for("_home"), code=302)
+
     log_file = request.args.get("log", 0, type=str)
     file_content = ClientLogReader(_DISPATCHER).get_client_log_content(client_id, log_file).replace("\n", "<br><br>") # pylint: disable=C0301
     return file_content
@@ -61,7 +80,15 @@ def _log_display():
 @_APP.route("/log_purge")
 def _log_purge():
     global _DISPATCHER # pylint: disable=W0603
+
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+
     client_id = request.args.get("client", 0, type=str)
+    if authenticated_client_id not in (Client.ROOT, client_id):
+        return redirect(url_for("_home"), code=302)
+
     _DISPATCHER.get_client_database(client_id).delete_old_logs(datetime.now())
     return redirect(url_for("_log_list"), code=302)
 
@@ -72,13 +99,29 @@ def _log_purge():
 @_APP.route("/queue_list")
 def _queue_list():
     global _DISPATCHER # pylint: disable=W0603
-    entries = ClientPassengerQueueReader(_DISPATCHER).get_client_passenger_queue_list()
+
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+    if authenticated_client_id == Client.ROOT:
+        authenticated_client_id = None
+
+    queue_reader = ClientPassengerQueueReader(_DISPATCHER)
+    entries = queue_reader.get_client_passenger_queue_list(p_client_id=authenticated_client_id)
     return render_template("queue_list.html", entries=entries)
 
 @_APP.route("/queue_display")
 def _queue_display():
     global _DISPATCHER # pylint: disable=W0603
+
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+
     client = request.args.get("client", 0, type=str)
+    if authenticated_client_id not in (Client.ROOT, client):
+        return redirect(url_for("_home"), code=302)
+
     passenger = request.args.get("passenger", 0, type=str)
     entry = ClientPassengerQueueReader(_DISPATCHER).get_client_passenger_queue_entry(client, passenger) # pylint: disable=C0301
     return render_template("queue_display.html", client=client, entry=entry)
@@ -86,7 +129,15 @@ def _queue_display():
 @_APP.route("/queue_attachment")
 def _queue_attachment():
     global _DISPATCHER # pylint: disable=W0603
+
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+
     client = request.args.get("client", 0, type=str)
+    if authenticated_client_id not in (Client.ROOT, client):
+        return redirect(url_for("_home"), code=302)
+
     passenger = request.args.get("passenger", 0, type=str)
     file = request.args.get("file", 0, type=str)
     entry = ClientPassengerQueueReader(_DISPATCHER).get_client_passenger_queue_entry(client, passenger) # pylint: disable=C0301
@@ -98,9 +149,17 @@ def _queue_attachment():
     return "File not found"
 
 @_APP.route("/queue_status_update")
-def _queue_status_update():
+def _queue_status_update(): # pylint: disable=R0912
     global _DISPATCHER # pylint: disable=W0603
-    client = request.args.get("client", "", type=str)
+
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+
+    client = request.args.get("client", 0, type=str)
+    if authenticated_client_id not in (Client.ROOT, client):
+        return redirect(url_for("_home"), code=302)
+
     passenger = request.args.get("passenger", "", type=str)
     entry = ClientPassengerQueueReader(_DISPATCHER).get_client_passenger_queue_entry(client, passenger) # pylint: disable=C0301
 
@@ -140,7 +199,15 @@ def _queue_status_update():
 @_APP.route("/queue_purge")
 def _queue_purge():
     global _DISPATCHER # pylint: disable=W0603
+
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+
     client_id = request.args.get("client", 0, type=str)
+    if authenticated_client_id not in (Client.ROOT, client_id):
+        return redirect(url_for("_home"), code=302)
+
     passenger_module = request.args.get("passenger", 0, type=str)
     client_db = _DISPATCHER.get_client_database(client_id)
     client_passenger = client_db.client.get_client_passenger(passenger_module)
@@ -160,7 +227,16 @@ def _queue_purge():
 @_APP.route("/passenger_list")
 def _passenger_list():
     global _DISPATCHER # pylint: disable=W0603
-    clients = _DISPATCHER.all_clients
+
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+
+    if authenticated_client_id == Client.ROOT:
+        clients = _DISPATCHER.all_clients
+    else:
+        clients = [_DISPATCHER.get_client_database(authenticated_client_id).client]
+
     try:
         expedited = request.args.get("expedited", 0, type=str) == "true"
     except Exception: # pylint: disable=W0703
@@ -170,25 +246,48 @@ def _passenger_list():
 @_APP.route("/passenger_expedite")
 def _passenger_expedite():
     global _DISPATCHER # pylint: disable=W0603
+
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+
     client = request.args.get("client", 0, type=str)
+    if authenticated_client_id not in (Client.ROOT, client):
+        return redirect(url_for("_home"), code=302)
+
     passenger = request.args.get("passenger", 0, type=str)
     _DISPATCHER.expedite_client_passenger(client, passenger)
     return redirect(url_for("_passenger_list") + "?expedited=true" + _get_cache_buster(), code=302)
 
 ##############################
-# Misc. pages
+# Peek
 ##############################
 
 @_APP.route("/peek")
 def _peek():
     global _DISPATCHER # pylint: disable=W0603
-    peek = PullerPeek(_DISPATCHER).peek()
+
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+    if authenticated_client_id == Client.ROOT:
+        authenticated_client_id = None
+
+    peek = PullerPeek(_DISPATCHER).peek(authenticated_client_id)
     return render_template("peek.html", peek=peek)
 
 @_APP.route("/peek_attachment")
 def _peek_attachment():
     global _DISPATCHER # pylint: disable=W0603
+
+    authenticated_client_id = _get_authenticated_client_id()
+    if authenticated_client_id is None:
+        return redirect(url_for("_login"), code=302)
+
     client = request.args.get("client", 0, type=str)
+    if authenticated_client_id not in (Client.ROOT, client):
+        return redirect(url_for("_home"), code=302)
+
     puller = request.args.get("puller", 0, type=str)
     passenger = request.args.get("passenger", 0, type=str)
     file_name = request.args.get("file", 0, type=str)
@@ -203,6 +302,43 @@ def _peek_attachment():
         return "File not found"
 
     return _download_attachment(attachment)
+
+##############################
+# Login
+##############################
+
+@_APP.route("/")
+def _login():
+    return render_template("login.html")
+
+@_APP.route("/login_attempt", methods=["POST"])
+def _login_attempt():
+    global _DISPATCHER # pylint: disable=W0603
+
+    try:
+        client_id = request.form["client_id"]
+        username = request.form["username"]
+        password = request.form["password"]
+
+        client_db = _DISPATCHER.get_client_database(client_id)
+        if client_db.client.authorization_active:
+            credential = Credential(username=username, password=password)
+            if client_db.client.authenticate(credential) is None:
+                return redirect(url_for("_login"), code=302)
+        session["client_id"] = client_db.client_id
+        return redirect(url_for("_home"), code=302)
+
+    except Exception: # pylint: disable=W0703
+        return redirect(url_for("_login"), code=302)
+
+@_APP.route("/logoff")
+def _logoff():
+    session["client_id"] = ""
+    return redirect(url_for("_login"), code=302)
+
+##############################
+# Misc
+##############################
 
 @_APP.route("/about")
 def _about():
@@ -233,3 +369,8 @@ def _download_attachment(att: Attachment):
 
 def _get_cache_buster() -> str:
     return "&cache_buster=" + str(uuid.uuid1())
+
+def _get_authenticated_client_id() -> str:
+    if "client_id" in session and session["client_id"] != "":
+        return session["client_id"]
+    return None
